tools/trace_plot/requirements.txt
pyusb
matplotlib
